Vastly simplified the subtype check.
Subtype checking where the subtype is an implicit positive
type map now just defers to check_expr() with an
inference context as the expression.
In retrospect, this is actually the obvious solution, but well.

Also, implicit type maps are now only emitted if their
potential binding is mentioned more than once in the
containing expression.